1738 Debugging dynamic code relative calls Do not try to parse code object name as filename if it doesn't look like one - <string> etc.
1739 [MMD] Cannot expand object with a null T_STRING slot
Treat null pointers in T_OBJECT and T_STRING slots as None.

1749 Navigate To shows files twice
.py files were incorrectly reported as Classes rather than Package to Object Browser.

1789 [MMD] Editing large int & long values crashes VS on x64
Fix copy/paste cast error.

1791 [MMD] Complex values don't show fractions correctly
Fix a long-standing bug in formatting string in GetConstantRepr for complex values.

1797 [MMD] CTypes stepping stops in _ctypes.c on x64
Don't skip module load checks if runtime is already loaded, so that there isn't a race between runtime creation and import ctypes.

1815 [MMD] Step In on next() does not step into tp_iternext
Fix off-by-one error in runtime exit point computation.

1824 Literals are not quoted in expressions in default argument values
Use GetConstantRepr instead of ToString.

1825 Floating-point literals use locale settings
Use InvariantCulture for double literals.

+++ Python/Product/Debugger/DkmDebugger/Proxies/Structs/PyLongObject.cs
@@ -68,7 +68,7 @@ namespace Microsoft.PythonTools.DkmDebugger.Proxies.Structs {
                 }
             } else {
                 for (var digitPtr = new UInt32Proxy(process, result.ob_digit.Address); absValue != 0; digitPtr = digitPtr.GetAdjacentProxy(1)) {
-                    digitPtr.Write((ushort)(absValue % (1 << bitsInDigit)));
+                    digitPtr.Write((uint)(absValue % (1 << bitsInDigit)));
                     absValue >>= bitsInDigit;
                 }
             }
